Only use suspend logic in app (#6212)
